% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{pca}
\alias{pca}
\title{Phylogenetic heatmap}
\usage{
pca(tree, Y, method = "BM", mode = "cov")
}
\arguments{
\item{tree}{tree}

\item{Y}{Y}

\item{method}{method}

\item{mode}{mode}
}
\description{
This function allows you to heatmap
}
\keyword{heatmap}
