% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{pca}
\alias{pca}
\title{Phylogenetic PCA}
\usage{
pca(tree, Y, method = "BM", mode = "cov")
}
\arguments{
\item{tree}{phylogeny as an object of class "phylo"}

\item{Y}{data matrix with traits in columns}

\item{method}{method to obtain the correlation structure: can be "BM" or "lambda"}

\item{mode}{can be "cov" or "corr"}
}
\description{
This function performs phylogenetic PCA
}
\keyword{PCA}
