% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloANOVA.R
\name{phyloANOVA}
\alias{phyloANOVA}
\title{Phylogenetic heatmap}
\usage{
phyloANOVA(tree, x, y, nsim = 1000, posthoc = TRUE, p.adj = "holm")
}
\arguments{
\item{tree}{tree}

\item{Y}{Y}

\item{method}{method}

\item{mode}{mode}
}
\description{
This function allows you to heatmap
}
\keyword{heatmap}
